pyproject.toml
[tool.poetry]
name = "carfigures"
version = "2.2.2"
description = "a fork of ballsdex that aims to provide a different experience."
authors = ["array_ye <array_ye@proton.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11, <3.14"

# asyncio
uvloop = { version = "^0.21.0", markers = "sys_platform != 'win32'" }

# Panel
fastapi = "^0.115.6"
fastapi-admin = {git = "https://github.com/Ballsdex-Team/fastapi-admin", branch = "ballsdex"}
uvicorn = "^0.34.0"

# TortoiseORM
tortoise-orm = {extras = ["asyncpg"], version = "^0.23.0"}
tortoise-cli = "^0.1.2"

rich = "^13.9.4"
python-dateutil = "^2.9.0"
Pillow = "^11.1.0"
aerich = "^0.6.3"
cachetools = "^5.5.2"
discord-py = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
ruff = "^0.9.7"

[tool.poetry.group.metrics.dependencies]
prometheus-client = "^0.16.0"

[tool.aerich]
tortoise_orm = "carfigures.__main__.TORTOISE_ORM"
location = "./migrations"
src_folder = "./carfigures"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 99

